Refactor builders into their own submodule
Mainly for tidyness of documentation. Some new trickiness is used to
make the inner json builders accessible to their use sites.
